Renamed `quran` app to `dquran`
