+++ bless/dquran/windows.py
